backend/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
resend
bcrypt
hypothesis